requirements.txt
python-dotenv
playwright
pymongo
redis
discord.py
